demo/wgpudemo/pyproject.toml
[tool.setuptools]
packages = ["wgpudemo"]

[project]
name = "wgpudemo"
version = "0.1.0"
description = 'Crunge WebGPU Demo'
requires-python = ">=3.8"
authors = [{ name = "Kurtis Fields", email = "kurtisfields@gmail.com" }]

dependencies = [
    "glfw ==2.7.0",
    "loguru ==0.7.2",
    "PyGLM ==2.7.1",
    "trimesh ==4.4.3",
    "imageio ==2.34.2",
    "imgui-bundle ==1.6.2",
]
